core: provide struct user_ta_store_ops in public .h file
Moves struct user_ta_store_ops definition into the new
<kernel/user_ta_store.h> file.
